Refactor billing import API test and extract test data
Moved billing import API test logic to a new test file and extracted request data to a shared test-data module for reuse. Also simplified the calculation of seconds in 24 hours in cloud accounts tests for clarity.
